requirements.txt
# Python 3.10+ required for MCP
mcp-python>=0.1.0
aiohttp>=3.8.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0.0
asyncio>=3.4.3
typing-extensions>=4.5.0
requests>=2.28.0
pytest>=7.3.1
python-dotenv>=1.0.0 